Refactor state management into EmailChangeForm
