+++ src/features/profile/DisplaySettings/DisplaySettings.test.tsx
@@ -3,9 +3,6 @@ import * as React from 'react';
 
 import { DisplaySettings } from './DisplaySettings';
 
-import Notice from 'src/components/Notice';
-
-
 describe('Email change form', () => {
   const update = jest.fn();
 
@@ -24,23 +21,5 @@ describe('Email change form', () => {
 
   it('should render', () => {
     expect(component).toHaveLength(1);
-  });
-
-  it('should display a notice on success.', () => {
-    const success = 'Account information updated.';
-    component.setState({ success });
-    expect(component.containsMatchingElement(<Notice success text={success} />)).toBeTruthy();
-  });
-
-  it('should display a notice for a general error', () => {
-    const errors = [{'reason': 'Something bad'}];
-    component.setState({ errors });
-    expect(component.containsMatchingElement(<Notice error text={'Something bad'} />)).toBeTruthy();
-  });
-
-  it('should not render the email form when loading', () => {
-    expect(component.find('[data-qa-email-change]')).toHaveLength(1);
-    component.setProps({ loading: true });
-    expect(component.find('[data-qa-email-change]')).toHaveLength(0);
   });
 });
